# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AI Podcast Web Admin Log API
  version: 1.0.0
  description: >
    日志管理相关接口规范（管理员配置日志开关、查询日志事件、前端全局异常上报）。
servers:
  - url: /
paths:
...
  /api/admin/log-events:
    get:
      summary: 查询日志事件与错误聚类
      parameters:
        - in: query
          name: scope
          schema:
            type: string
            enum: [notebook_share_client, frontend_global_error]
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - in: query
          name: level
          schema:
            type: string
            enum: [info, error]
        - in: query
          name: requestId
          schema:
            type: string
        - in: query
          name: errorCode
          schema:
            type: string
        - in: query
          name: fromMs
          schema:
            type: integer
        - in: query
          name: toMs
          schema:
            type: integer
      responses:
        "200":
          description: 成功
        "401":
          description: 未登录
        "403":
          description: 非管理员
...